requirements.txt
Django==5.2.5
djangorestframework==3.16.1
requests==2.32.5
python-decouple==3.8
Pillow==11.3.0

